Simplify Money Project/backend/requirements.txt
flask
flask-cors
python-dotenv
openai

